libtsol: comparison of integers of different signs
/code/illumos-gate/usr/src/common/tsol/stol.c:138:15: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
            (bytes*2 != strlen(h)) ||
             ~~~~~~~ ^  ~~~~~~~~~
/code/illumos-gate/usr/src/common/tsol/stol.c:137:13: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        if ((bytes > len) ||
             ~~~~~ ^ ~~~
2 errors generated.

../common/zone.c:123:16: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        for (i = 0; i < nzents; i++) {
                    ~ ^ ~~~~~~
../common/zone.c:238:16: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        for (i = 0; i < nzents; i++) {
                    ~ ^ ~~~~~~
2 errors generated.

../common/btos.c:67:21: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
        } else if (val_len > str_len) {
                   ~~~~~~~ ^ ~~~~~~~
../common/btos.c:360:14: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        if (wccount > len) {
            ~~~~~~~ ^ ~~~
../common/btos.c:456:14: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        if (wccount > len) {
            ~~~~~~~ ^ ~~~
3 errors generated.

+++ usr/src/common/tsol/stol.c
@@ -107,7 +107,7 @@ htol(const char *s, m_label_t *l)
 	const char	*h = &s[2];	/* skip 0[xX] */
 	uchar_t *lp = (uchar_t *)&(((_mac_label_impl_t *)l)->_lclass);
 	size_t	len = sizeof (_mac_label_impl_t) - 4;
-	int	bytes;
+	size_t	bytes;
 
 	/* unpack 16 bit signed classification */
 	if (!unhex(&h, lp, 2) || (LCLASS(l) < 0)) {
@@ -133,9 +133,9 @@ htol(const char *s, m_label_t *l)
 		bzero(lp, len);
 		h++;	/* skip '-' */
 	}
-	bytes = strlen(h)/2;
+	bytes = strlen(h) / 2;
 	if ((bytes > len) ||
-	    (bytes*2 != strlen(h)) ||
+	    (bytes * 2 != strlen(h)) ||
 	    !unhex(&h, lp, bytes)) {
 		return (-1);
 	}
